breakpoint occuring in FlamethrowerPowerUp.cpp
added flamethrower sounds
couldn't figure out where to put the "ChangeWeapons" audio in C++ (in equipbasicweapon for FlightPlayer.cpp)
couldn't find the code for audio for pick-ups (got me so confused) to put for nuke, shield and basic weapon.

PS: thanks for putting the audios in c++ earlier. Didn't even try to touch the other code. Dont want to ruin it.